+++ amelie/repl/repl/streamer.c
@@ -204,7 +204,7 @@ streamer_process(Streamer* self)
 		cancellation_point();
 
 		// reconnect
-		auto reconnect_interval = opt_int_of(&config()->repl_reconnect_ms);
+		auto reconnect_interval = (int)opt_int_of(&config()->repl_reconnect_ms);
 		info("reconnect in %d ms", reconnect_interval);
 		coroutine_sleep(reconnect_interval);
 	}
